make the old test keras-only
