+++ e2e_tests/tests/experiment/experiment.py
@@ -1,6 +1,5 @@
 import datetime
 import logging
-import multiprocessing
 import os
 import re
 import subprocess
@@ -13,7 +12,6 @@ import dateutil.parser
 import pytest
 import requests
 
-from determined import experimental
 from determined.common import api, yaml
 from determined.common.api import authentication, certs
 from tests import config as conf
@@ -700,23 +698,3 @@ def s3_checkpoint_config_no_creds() -> Dict[str, str]:
 
 def root_user_home_bind_mount() -> Dict[str, str]:
     return {"host_path": "/tmp", "container_path": "/root"}
-
-
-def _export_and_load_model(experiment_id: int, master_url: str) -> None:
-    experimental.Determined(master_url).get_experiment(experiment_id).top_checkpoint().load()
-
-
-def export_and_load_model(experiment_id: int) -> None:
-    # We run this in a subprocess to avoid module name collisions
-    # when performing checkpoint export of different models.
-    ctx = multiprocessing.get_context("spawn")
-    p = ctx.Process(
-        target=_export_and_load_model,
-        args=(
-            experiment_id,
-            conf.make_master_url(),
-        ),
-    )
-    p.start()
-    p.join()
-    assert p.exitcode == 0, p.exitcode
